Fix AMP dtype crash in the loss on the first batch
Under autocast the head emits half, but ground truth arrives from the DataLoader
as float32. The assigner's IoU therefore promotes to float32, and writing that
result back into the half classification-target tensor raises

    RuntimeError: Index put requires the source and destination dtypes match,
    got Half for the destination and Float for the source

This is a first-batch crash: AMP training could never start.

Fixed at the root by computing the loss in fp32 -- the standard AMP arrangement,
since QFL, DFL and GIoU are all numerically better behaved there. The token
consistency loss is cast too: a spatial softmax over H*W positions underflows
badly in half.

tests/test_amp_loss.py reproduces the promotion path with CPU autocast
(bfloat16), so it catches this without a GPU, and also asserts every loss term
comes back fp32 and that a half model with float32 targets does not raise.

+++ losses/token_consistency.py
@@ -30,7 +30,9 @@ class TokenConsistencyLoss(nn.Module):
                 teacher_maps: dict[str, torch.Tensor]) -> torch.Tensor:
         losses = []
         for lv, s in student_maps.items():
-            t = teacher_maps[lv].detach()
+            # fp32: a spatial softmax over H*W positions underflows badly in half
+            s = s.float()
+            t = teacher_maps[lv].detach().float()
             b = s.shape[0]
             s_flat = s.reshape(b, -1) / self.t
             t_flat = t.reshape(b, -1) / self.t
